Revert "feat: Grid component"
