+++ src/storybook/stand-alone-documentaion/grid/grid.mdx
@@ -1,7 +1,7 @@
 import { Meta, Story, Canvas } from "@storybook/blocks";
 import * as GridStories from "./Grid.stories";
 
-<Meta title="Layout/Grid" />
+<Meta title="Foundations/Grid" />
 
 ## Basic usage
 
@@ -20,24 +20,18 @@ Breakpoints size: **xs**: 1rem, **sm**: 48rem, **md**: 64rem, **lg**: 90rem, **x
 > for a custom configuration, [see this section](/custom)
 
 ```js
-import { GridCol, GridContainer, GridRow } from "../Grid";
+import { Container, Row, Col } from "react-awesome-styled-grid";
 
-const JustifyContent = {
-  render: () => {
-    return (
-      <GridContainer>
-        <GridRow>
-          <GridCol xs={4} md={2}>
-            Col A
-          </GridCol>
-          <GridCol xs={4} md={6}>
-            Col B
-          </GridCol>
-        </GridRow>
-      </GridContainer>
-    );
-  },
-
-  name: "justifyContent"
-};
+const MyCoolComponent = () => (
+  <Container>
+    <Row>
+      <Col xs={4} md={2}>
+        Col A
+      </Col>
+      <Col xs={4} md={6}>
+        Col B
+      </Col>
+    </Row>
+  </Container>
+);
 ```
